:recycle: Refactor IR type name module
