+++ woke/ast/ir/abc.py
@@ -4,17 +4,7 @@ from typing import Optional, Tuple
 
 from woke.ast.ir.reference_resolver import ReferenceResolver
 from woke.ast.ir.utils.init_tuple import IrInitTuple
-from woke.ast.nodes import (
-    SolcArrayTypeName,
-    SolcElementaryTypeName,
-    SolcExpressionUnion,
-    SolcFunctionTypeName,
-    SolcMapping,
-    SolcNode,
-    SolcStatementUnion,
-    SolcTypeNameUnion,
-    SolcUserDefinedTypeName,
-)
+from woke.ast.nodes import SolcExpressionUnion, SolcNode, SolcStatementUnion
 
 
 class IrAbc(ABC):
@@ -61,45 +51,6 @@ class IrAbc(ABC):
             self._ast_node.src.byte_offset,
             self._ast_node.src.byte_offset + self._ast_node.src.byte_length,
         )
-
-
-class TypeNameAbc(IrAbc):
-    _type_identifier: Optional[str]
-    _type_string: Optional[str]
-
-    def __init__(self, init: IrInitTuple, type_name: SolcTypeNameUnion, parent: IrAbc):
-        super().__init__(init, type_name, parent)
-        self._type_identifier = type_name.type_descriptions.type_identifier
-        self._type_string = type_name.type_descriptions.type_string
-
-    @staticmethod
-    def from_ast(
-        init: IrInitTuple, type_name: SolcTypeNameUnion, parent: IrAbc
-    ) -> "TypeNameAbc":
-        from woke.ast.ir.type_name.array_type_name import ArrayTypeName
-        from woke.ast.ir.type_name.elementary_type_name import ElementaryTypeName
-        from woke.ast.ir.type_name.function_type_name import FunctionTypeName
-        from woke.ast.ir.type_name.mapping import Mapping
-        from woke.ast.ir.type_name.user_defined_type_name import UserDefinedTypeName
-
-        if isinstance(type_name, SolcArrayTypeName):
-            return ArrayTypeName(init, type_name, parent)
-        elif isinstance(type_name, SolcElementaryTypeName):
-            return ElementaryTypeName(init, type_name, parent)
-        elif isinstance(type_name, SolcFunctionTypeName):
-            return FunctionTypeName(init, type_name, parent)
-        elif isinstance(type_name, SolcMapping):
-            return Mapping(init, type_name, parent)
-        elif isinstance(type_name, SolcUserDefinedTypeName):
-            return UserDefinedTypeName(init, type_name, parent)
-
-    @property
-    def type_identifier(self) -> Optional[str]:
-        return self._type_identifier
-
-    @property
-    def type_string(self) -> Optional[str]:
-        return self._type_string
 
 
 class ExpressionAbc(IrAbc):
