+++ woke/ast/ir/meta/using_for_directive.py
@@ -3,13 +3,14 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Union
 
 from woke.ast.ir.meta.identifier_path import IdentifierPath
+from woke.ast.ir.type_name.abc import TypeNameAbc
 from woke.ast.ir.type_name.user_defined_type_name import UserDefinedTypeName
 from woke.ast.ir.utils import IrInitTuple
 
 if TYPE_CHECKING:
-    from woke.ast.ir.declaration import ContractDefinition
+    from woke.ast.ir.declaration.contract_definition import ContractDefinition
 
-from woke.ast.ir.abc import IrAbc, TypeNameAbc
+from woke.ast.ir.abc import IrAbc
 from woke.ast.nodes import (
     SolcIdentifierPath,
     SolcUserDefinedTypeName,
