+++ woke/ast/ir/type_name/array_type_name.py
@@ -1,6 +1,7 @@
 from typing import Optional, Union
 
-from woke.ast.ir.abc import ExpressionAbc, IrAbc, TypeNameAbc
+from woke.ast.ir.abc import ExpressionAbc, IrAbc
+from woke.ast.ir.type_name.abc import TypeNameAbc
 from woke.ast.ir.utils import IrInitTuple
 from woke.ast.nodes import SolcArrayTypeName
