+++ woke/ast/ir/type_name/function_type_name.py
@@ -1,6 +1,7 @@
 from woke.ast.enums import StateMutability, Visibility
-from woke.ast.ir.abc import IrAbc, TypeNameAbc
+from woke.ast.ir.abc import IrAbc
 from woke.ast.ir.meta.parameter_list import ParameterList
+from woke.ast.ir.type_name.abc import TypeNameAbc
 from woke.ast.ir.utils import IrInitTuple
 from woke.ast.nodes import SolcFunctionTypeName
