+++ woke/ast/ir/type_name/mapping.py
@@ -1,4 +1,5 @@
-from woke.ast.ir.abc import IrAbc, TypeNameAbc
+from woke.ast.ir.abc import IrAbc
+from woke.ast.ir.type_name.abc import TypeNameAbc
 from woke.ast.ir.utils import IrInitTuple
 from woke.ast.nodes import SolcMapping
